Move /src/lib to /src, ignore __tests__ in build
