+++ packages/svelte-query/src/QueryClientProvider.svelte
