+++ packages/svelte-query/src/createInfiniteQuery.ts
