+++ packages/svelte-query/src/types.ts
