+++ packages/svelte-query/svelte.config.js
@@ -3,9 +3,12 @@ import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 const config = {
   preprocess: vitePreprocess(),
   package: {
-    source: "./src/lib",
-    dir: "./build/lib"
-  }
+    source: "./src",
+    dir: "./build/lib",
+    files: (filepath) => {
+      return !filepath.includes("__tests__");
+    },
+  },
 };
 
 export default config;
